+++ commitlint.config.js
@@ -1,3 +1,24 @@
+// eslint-disable-next-line no-undef
 module.exports = {
   extends: ['@commitlint/config-conventional'],
+  rules: {
+    'type-enum': [
+      2,
+      'always',
+      [
+        'build',
+        'chore',
+        'ci',
+        'docs',
+        'feat',
+        'fix',
+        'perf',
+        'refactor',
+        'revert',
+        'style',
+        'test',
+        'deprecate',
+      ],
+    ],
+  },
 };