+++ e2e/navigate-long-message-lists.test.ts
@@ -1,14 +1,21 @@
-import { expect, Page } from '@playwright/test';
-import dotenv from 'dotenv';
+/* eslint-disable jest/expect-expect */
+import { Page } from '@playwright/test';
+import * as dotenv from 'dotenv';
 
 import selectors from './user/selectors';
 import { CustomTestContext, test } from './user/test';
 
-import MessageInput from './user/components/MessageInput';
-import Thread from './user/components/Thread/Thread';
 import ChannelPreview from './user/components/ChannelPreview';
 import Message from './user/components/Message/MessageSimple';
+import MessageInput from './user/components/MessageInput';
+import MessageList from './user/components/MessageList/MessageList';
+import MessageNotification, {
+  getMessageNotificationSelector,
+} from './user/components/MessageList/MessageNotification';
 import QuotedMessage from './user/components/Message/QuotedMessage';
+import Thread, { composeThreadSelector } from './user/components/Thread/Thread';
+
+import type { TestingUser } from './user/User';
 
 dotenv.config();
 dotenv.config({ path: `.env.local` });
@@ -17,7 +24,9 @@ const user1Id = process.env.E2E_TEST_USER_1;
 const CHANNEL_NAME = 'navigate-long-message-lists' as const;
 const MY_ADDED_REPLY_TEXT = 'My reply' as const;
 const OTHER_USER_ADDED_REPLY_TEXT = 'Reply back' as const;
+const OTHER_USER_ADDED_MESSAGE_TEXT = "Other user's message" as const;
 const USER1_CHAT_VIEW_CLASSNAME = `.${user1Id}`;
+const NEW_MESSAGE_NOTIFICATION_TEXT = 'New Messages!' as const;
 const LAST_REPLY_TEXT = 'Message 299';
 const MESSAGES_WITH_REPLIES = ['Message 149', 'Message 137', 'Message 124', 'Message 99'];
 
@@ -84,7 +93,7 @@ test.describe('thread autoscroll', () => {
       await user.clicks(ChannelPreview).text(CHANNEL_NAME);
       await Promise.all([
         page.waitForResponse((r) => r.url().includes('/replies') && r.ok()),
-        user.clicks(Thread).openFor(MESSAGES_WITH_REPLIES[0]),
+        user.clicks(Thread).open('replies', -1),
       ]);
     });
 
@@ -103,37 +112,102 @@ test.describe('thread autoscroll', () => {
     });
 
     test('only if I send a message', async ({ page, user }) => {
-      let thread = await user.get(Thread)(USER1_CHAT_VIEW_CLASSNAME);
-      const avatars = await thread.locator(selectors.avatar);
-      const message = await user.get(Message)('Message 270');
-      await message.scrollIntoViewIfNeeded();
-      await message.waitFor({ state: 'visible', timeout: 3000 });
+      const selector = composeThreadSelector(USER1_CHAT_VIEW_CLASSNAME);
+      await user.sees(Thread).not.isScrolledToBottom(selector);
 
       await Promise.all([
         page.waitForResponse((r) => r.url().includes('/message') && r.ok()),
         user.submits(MessageInput).reply(MY_ADDED_REPLY_TEXT),
       ]);
 
-      await expect(thread).toHaveScreenshot({
-        mask: [avatars],
-      });
+      await user.sees(Thread).isScrolledToBottom(selector);
     });
 
     test('not if I receive a message', async ({ controller, page, user }) => {
-      const thread = await user.get(Thread)(USER1_CHAT_VIEW_CLASSNAME);
-      const avatars = await thread.locator(selectors.avatar);
-      const message = await user.get(Message)('Message 270');
-      await message.scrollIntoViewIfNeeded();
-      await message.waitFor({ state: 'visible', timeout: 3000 });
+      const selector = composeThreadSelector(USER1_CHAT_VIEW_CLASSNAME);
+      await user.sees(Thread).not.isScrolledToBottom(selector);
 
       await Promise.all([
         page.waitForResponse((r) => r.url().includes('/message') && r.ok()),
         await controller.sendOtherUserReply(),
       ]);
-
-      await expect(thread).toHaveScreenshot({
-        mask: [avatars],
-      });
+      await user.sees(Thread).not.isScrolledToBottom(selector);
     });
+  });
+});
+
+test.describe('scroll to the bottom', () => {
+  const scrollInSteps = async (user: TestingUser, cycles = 1) => {
+    for (let i = 0; i < cycles; i++) {
+      await Promise.all(
+        ['142', '135', '128'].map((num: string) =>
+          user.get(Message)(`Message ${num}`).scrollIntoViewIfNeeded(),
+        ),
+      );
+    }
+  };
+  test.beforeEach(async ({ controller, user }) => {
+    await controller.openStory(
+      'navigate-long-message-lists--user1',
+      selectors.channelPreviewButton,
+    );
+    await user.clicks(ChannelPreview).text(CHANNEL_NAME);
+  });
+
+  test.afterEach(async ({ controller, page }) => {
+    const lastMessage = await page
+      .locator(
+        `${USER1_CHAT_VIEW_CLASSNAME} ${selectors.messageList} li:last-of-type ${selectors.messageText}`,
+      )
+      .textContent();
+    if (!lastMessage) return;
+    if (lastMessage.match(OTHER_USER_ADDED_MESSAGE_TEXT)) {
+      await controller.deleteOtherUserLastMessage();
+    }
+  });
+
+  test('without loading more messages on new message notification click', async ({
+    controller,
+    page,
+    user,
+  }) => {
+    // scroll without loading more messages
+    await scrollInSteps(user, 3);
+
+    await controller.sendOtherUserMessage();
+
+    // click the notification
+    await page.waitForSelector(getMessageNotificationSelector(NEW_MESSAGE_NOTIFICATION_TEXT));
+    await user.clicks(MessageNotification).text(NEW_MESSAGE_NOTIFICATION_TEXT);
+
+    // check that you are at the bottom
+    await user
+      .sees(MessageList)
+      .isScrolledToBottom(`${USER1_CHAT_VIEW_CLASSNAME} ${selectors.messageList}`);
+  });
+
+  test('after loading more messages on new message notification click', async ({
+    controller,
+    page,
+    user,
+  }) => {
+    // scroll without loading more messages
+    await scrollInSteps(user, 3);
+
+    // trigger load more messages
+    const firstLoadedMessage = await page.locator(
+      `${USER1_CHAT_VIEW_CLASSNAME} ${selectors.messageList} li:first-of-type`,
+    );
+    await firstLoadedMessage.scrollIntoViewIfNeeded();
+    await controller.sendOtherUserMessage();
+
+    // click the notification
+    await page.waitForSelector(getMessageNotificationSelector(NEW_MESSAGE_NOTIFICATION_TEXT));
+    await user.clicks(MessageNotification).text(NEW_MESSAGE_NOTIFICATION_TEXT);
+
+    // check that you are at the bottom
+    await user
+      .sees(MessageList)
+      .isScrolledToBottom(`${USER1_CHAT_VIEW_CLASSNAME} ${selectors.messageList}`);
   });
 });