+++ e2e/user/components/Thread/Thread.ts
@@ -1,14 +1,24 @@
+/* eslint-disable jest/no-standalone-expect */
 import type { Page } from '@playwright/test';
 import selectors from '../../selectors';
 import { expect } from '@playwright/test';
 import { getMessage } from '../Message/MessageSimple';
+
+export function composeThreadSelector(prependSelectors?: string) {
+  return `${prependSelectors || ''} ${selectors.threadMessageList}`;
+}
 
 export default (page: Page) => ({
   click: {
     close() {
       return page.click(selectors.buttonCloseThread);
     },
-    async openFor(messageText: string, nth: number = 0) {
+    open(text: string, nth = 0) {
+      return page
+        .locator(`${selectors.messageRepliesButton} >> nth=${nth} `, { hasText: text })
+        .click();
+    },
+    async openFor(messageText: string, nth = 0) {
       await page
         .locator(selectors.message, { hasText: messageText })
         .locator(selectors.messageRepliesButton)
@@ -16,8 +26,7 @@ export default (page: Page) => ({
         .click();
     },
   },
-  get: (prependSelectors?: string) =>
-    page.locator(`${prependSelectors || ''} ${selectors.threadMessageList}`),
+  get: (prependSelectors?: string) => page.locator(composeThreadSelector(prependSelectors)),
   see: {
     empty() {
       const replies = page.locator(`${selectors.threadReplyList} ${selectors.message}`);
@@ -34,6 +43,18 @@ export default (page: Page) => ({
         // const msgBox = await getMessage(page, text, nth, isThread).boundingBox();
         // expect(isVisible(msgBox, threadMsgListBox)).toBeTruthy();
       },
+    },
+    async isScrolledToBottom(selector: string) {
+      expect(
+        await page.evaluate(
+          ([selector]) => {
+            const messageList = document.querySelector(selector);
+            if (!messageList) return false;
+            return messageList.scrollTop + messageList.clientHeight === messageList.scrollHeight;
+          },
+          [selector],
+        ),
+      ).toBeTruthy();
     },
     not: {
       empty() {
@@ -61,6 +82,18 @@ export default (page: Page) => ({
           // const msgBox = await msgLocator.boundingBox();
           // expect(isVisible(msgBox, threadMsgListBox)).toBeFalsy();
         },
+      },
+      async isScrolledToBottom(selector: string) {
+        expect(
+          await page.evaluate(
+            ([selector]) => {
+              const messageList = document.querySelector(selector);
+              if (!messageList) return false;
+              return messageList.scrollTop + messageList.clientHeight === messageList.scrollHeight;
+            },
+            [selector],
+          ),
+        ).toBeFalsy();
       },
     },
     start() {