+++ src/components/Channel/channelState.ts
@@ -177,6 +177,7 @@ export const channelReducer = <
         hasMore,
         loadingMore: false,
         messages,
+        suppressAutoscroll: false,
       };
     }
 
@@ -218,7 +219,7 @@ export const channelReducer = <
     case 'setLoadingMore': {
       const { loadingMore } = action;
       // suppress the autoscroll behavior
-      return { ...state, loadingMore, suppressAutoscroll: true };
+      return { ...state, loadingMore, suppressAutoscroll: loadingMore };
     }
 
     case 'setLoadingMoreNewer': {