+++ src/components/Message/MessageStatus.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { DeliveredCheckIcon, MessageDeliveredIcon } from './icons';
-import { getReadByTooltipText } from './utils';
+import { getReadByTooltipText, mapToUserNameOrId, TooltipUsernameMapper } from './utils';
 
 import { AvatarProps, Avatar as DefaultAvatar } from '../Avatar';
 import { LoadingIndicator } from '../Loading';
@@ -17,6 +17,7 @@ import type { DefaultStreamChatGenerics } from '../../types/types';
 export type MessageStatusProps = {
   Avatar?: React.ComponentType<AvatarProps>;
   messageType?: string;
+  tooltipUserNameMapper?: TooltipUsernameMapper;
 };
 
 const UnMemoizedMessageStatus = <
@@ -24,7 +25,11 @@ const UnMemoizedMessageStatus = <
 >(
   props: MessageStatusProps,
 ) => {
-  const { Avatar: propAvatar, messageType = 'simple' } = props;
+  const {
+    Avatar: propAvatar,
+    messageType = 'simple',
+    tooltipUserNameMapper = mapToUserNameOrId,
+  } = props;
 
   const { client } = useChatContext<StreamChatGenerics>('MessageStatus');
   const { Avatar: contextAvatar } = useComponentContext<StreamChatGenerics>('MessageStatus');
@@ -61,7 +66,7 @@ const UnMemoizedMessageStatus = <
 
     return (
       <span className={rootClassName} data-testid='message-status-read-by'>
-        <Tooltip>{getReadByTooltipText(readBy, t, client)}</Tooltip>
+        <Tooltip>{getReadByTooltipText(readBy, t, client, tooltipUserNameMapper)}</Tooltip>
         <Avatar
           image={lastReadUser.image}
           name={lastReadUser.name || lastReadUser.id}